Refactor repository module.
